refactor: move exit-gate blocking detection to internal/review
The merged-report "is this blocking?" heuristic lived in the danger-zone
runner.go as four pure string functions (isBlockingMarkdown +
recommendationIsBlocking + sectionHasContent + isNonePlaceholder). It's
contract-critical — it decides exit 0 vs exit 2, which pre-commit hooks
act on — yet it sat in the command package where it's awkward to test in
isolation.

Moves it to internal/review as the exported review.IsBlockingMarkdown
(helpers stay unexported). review only imports git, so no cycle. The
orchestration that depends on multi.ReviewResult (decideExitGate,
anyPerLLMHasBlocking) stays in cmd and now calls review.IsBlockingMarkdown.

Behavior-preserving move: the table tests + real-report fixture moved with
it (TestIsBlockingMarkdown*), and the in-cmd TestDecideExitGate /
TestAnyPerLLMHasBlocking still pass through the new call path. runner.go
−114 LOC; internal/review coverage 90.4% → 93.0%.
